requirements.txt
flask==3.0.3
pillow==10.4.0
opencv-python==4.10.0.84
torch>=2.1.0
torchvision>=0.16.0
realesrgan==0.3.0